Enable video streaming with AP1302 ISP and AR1335 over MIPI CSI Interface (#13)
This commit enables video streaming over MIPI using our external ISP (AP1302) and sensor (AR1335).
A few fixes have been done to get this to work.

For example, the GPIO's used to control the AP1302 are on a IO Expander which is controlled via I2C. So all the GPIO function calls in the AP1302 driver may sleep when called as they will trigger an I2C transaction. The kernel provides a warning if you try to call gpiod_set_value() instead of gpiod_set_value_cansleep() for GPIO's on the IO expander. 

The RZG2L driver had to be modified slightly to accept our input format. And the AP1302 driver need an extra V4L2 control handler to handle pixel rate requests.